Added main communicator script for our modules
This is a very rough copy, assumed *a lot* about what the input/output of the other modules will be (yet TBD). The current version will not compile, but shows how the main thing is going to be structured. In addition, see top comment block for main explanation of our state machine.